17049 - Supervisors and featured defects

+++ js/sync.js
@@ -272,6 +272,9 @@ function Sync()
         self.syncingCounter++;
         self.syncingTotalRequest++;
 
+        if (tableName == 'builders_supervisors')
+            objDBUtils.emptyTable('builders_supervisors');
+		
         if(!self.silentMode) $("#accountMessage #general").text("Processing: " + (self.syncingTotalRequest - self.syncingCounter) + '/' + self.syncingTotalRequest);
         if(!self.silentMode) blockElement("body");
         $.post(objApp.apiURL + 'account/get_data_table/' + tableName +'/' + refreshSync, parameters , function(data)
@@ -696,7 +699,10 @@ function Sync()
         self.saveData.push('0'); /* dirty = 0 */
 		sql += header + ", dirty) " + footer + ", ?);";
 
-		
+		if (tableName == 'builders_supervisors'){
+		    console.log(sql);
+            console.log(self.saveData);
+        }
 		return sql;		
 	};	
 	
@@ -712,7 +718,6 @@ function Sync()
 
 			// Set all dirty records as not dirty
 			var sql = "UPDATE " + tableName + " SET dirty = 0 WHERE dirty = 1";
-			
 			if(tableName == "inspections")
 			{
 				// For the inspections table, only set dirty to 0 when the inspection has been finalised.
@@ -723,7 +728,7 @@ function Sync()
 
 			// Remove deleted records from local storage.  However, dont delete deleted contactfavourite records
             // as the deleted flag in this table simply means that the contact is not infact a favourite of this user.
-            if(tableName != "contactsfavourites") {
+            if(tableName != "contactsfavourites" && tableName != "builders_supervisors") {
 			    var sql = "DELETE FROM " + tableName + " WHERE deleted = 1;";
 			    objDBUtils.execute(sql, null, null);			
             }
